Add tool & workflow to fetch changelog
Rewrote the changelog script to use primarily merged pull requests
between "commit A" and "commit B".
It uses the GitHub rest api. The resulting data is pushed to a draft gist.

Manual usage is in the README.

Also adds a GitHub workflow action to run the tool on demand, through
the GitHub action's interface.