Add tool to bump version & create PR
This includes a tool to update the various files in the repo that need
to change when a version increments.

It also includes a workflow to trigger the action manually, passing it a
version & a changelog url. It'll update the values, build the app, build
the doc, make a commit in a branch, and create a pull request to merge
this stuff back into dev.